requirements.txt
pytest==7.1.2
requests==2.27.1
selenium==4.1.3
webdriver-manager-3.5.4
faker==13.11.1
